Accumulated changes
I took the liberty and gathered some of the PRs and forks from the original module (fbrnc/aoe_quotecleaner) and mashed them all together in on PR as I saw fit.

I really hope this gets merged, especially the no log on no removed records that otherwise is polluting the `system.log` file.